requirements.txt
# requirements.txt
torch==2.0.0+cu118
torchvision~=0.15.2
pandas~=2.0.1
numpy~=1.24.3
scikit-learn~=1.6.1
matplotlib~=3.7.1
seaborn~=0.12.2
Pillow~=10.3.0
timm==0.9.2
facenet-pytorch==2.5.3
tqdm==4.66.3
tensorboard==2.13.0
onnxruntime-gpu==1.15.1
onnx==1.14.0

